PF-577 Renamed the app and added a Readme
Renamed project to importer instead of Synchronizer: this is not a Service
  although it can be scheduled to run automatically as a synchronizer.
Also set up Appveyor to put the project into build and release.
