+++ Samples/DotNetSdk/DotNetSdkExamples.sln
@@ -7,7 +7,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ObservationValidator", "Obs
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SamplesPlannedSpecimenInstantiator", "SamplesPlannedSpecimenInstantiator\SamplesPlannedSpecimenInstantiator.csproj", "{C491945B-F4FD-4501-B8D6-B1CF671762F5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SondeFileSynchronizer", "SondeFileSynchronizer\SondeFileSynchronizer.csproj", "{6D5F7772-8BD0-4FCB-B2AC-0A173CD0C2E8}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SondeFileImporter", "SondeFileImporter\SondeFileImporter.csproj", "{6D5F7772-8BD0-4FCB-B2AC-0A173CD0C2E8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
