+++ Samples/DotNetSdk/SondeFileImporter/App.config
@@ -32,7 +32,7 @@
       </layout>
     </appender>
     <appender name="RollingFileAppender" type="log4net.Appender.RollingFileAppender">
-      <file value="Logs\SondeFileSynchronizer.log" />
+      <file value="Logs\SondeFileImporter.log" />
       <appendToFile value="true" />
       <rollingStyle value="Size" />
       <maxSizeRollBackups value="30" />
