+++ Samples/DotNetSdk/SondeFileImporter/Config.ini
@@ -1,7 +1,7 @@
 ﻿#General settings:
 ##All folder settings are optional. By default, the folders will be created under the tool's folder. 
-#SondeFileFolder=D:\SondeFileSynchronizer\SondeFiles
-#ArchiveFolder=D:\SondeFileSynchronizer\Archives
+#SondeFileFolder=D:\SondeFileImporter\SondeFiles
+#ArchiveFolder=D:\SondeFileImporter\Archives
 
 ##Samples API URL and auth token are required:
 SamplesApiBaseUrl=https://[your_account].aqsamples.com/api/
