+++ Samples/DotNetSdk/SondeFileImporter/Config/ConfigLoader.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 
-namespace SondeFileSynchronizer.Config
+namespace SondeFileImporter.Config
 {
     public class ConfigLoader
     {
