+++ Samples/DotNetSdk/SondeFileImporter/Config/Context.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 
-namespace SondeFileSynchronizer.Config
+namespace SondeFileImporter.Config
 {
     public class Context
     {
