+++ Samples/DotNetSdk/SondeFileImporter/ExampleDataFiles/Sonde file example.Converted.csv
@@ -1,0 +1,49 @@
+Observation ID,Location ID,Observed Property ID,Observed DateTime,Analyzed DateTime,Depth,Depth Unit,Data Classification,Result Value,Result Unit,Result Status,Result Grade,Medium,Sample ID,Collection Method,Field: Device ID,Field: Device Type,Field: Comment,Lab: Specimen Name,Lab: Analysis Method,Lab: Detection Condition,Lab: Limit Type,Lab: MDL,Lab: MRL,Lab: Quality Flag,Lab: Received DateTime,Lab: Prepared DateTime,Lab: Sample Fraction,Lab: From Laboratory,Lab: Sample ID,Lab: Dilution Factor,Lab: Comment,QC: Type,QC: Source Sample ID
+,Cn,Barometric Pressure,5/20/2019 8:30:08,,0.261,m,VERTICAL_PROFILE,760.4,mmHg,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Concentration),5/20/2019 8:30:08,,0.261,m,VERTICAL_PROFILE,4.07,mg/l,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Saturation),5/20/2019 8:30:08,,0.261,m,VERTICAL_PROFILE,51.6,%,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,pH,5/20/2019 8:30:08,,0.261,m,VERTICAL_PROFILE,7.25,s.u.,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Specific Conductivity,5/20/2019 8:30:08,,0.261,m,VERTICAL_PROFILE,124,uS/cm,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Water Temperature_C,5/20/2019 8:30:08,,0.261,m,VERTICAL_PROFILE,27.57,degC,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Barometric Pressure,5/20/2019 8:31:07,,0.497,m,VERTICAL_PROFILE,760.5,mmHg,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Concentration),5/20/2019 8:31:07,,0.497,m,VERTICAL_PROFILE,3.97,mg/l,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Saturation),5/20/2019 8:31:07,,0.497,m,VERTICAL_PROFILE,50.3,%,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,pH,5/20/2019 8:31:07,,0.497,m,VERTICAL_PROFILE,7.03,s.u.,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Specific Conductivity,5/20/2019 8:31:07,,0.497,m,VERTICAL_PROFILE,124,uS/cm,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Water Temperature_C,5/20/2019 8:31:07,,0.497,m,VERTICAL_PROFILE,27.54,degC,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Barometric Pressure,5/20/2019 8:32:02,,1,m,VERTICAL_PROFILE,760.4,mmHg,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Concentration),5/20/2019 8:32:02,,1,m,VERTICAL_PROFILE,3.65,mg/l,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Saturation),5/20/2019 8:32:02,,1,m,VERTICAL_PROFILE,46.2,%,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,pH,5/20/2019 8:32:02,,1,m,VERTICAL_PROFILE,6.83,s.u.,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Specific Conductivity,5/20/2019 8:32:02,,1,m,VERTICAL_PROFILE,124,uS/cm,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Water Temperature_C,5/20/2019 8:32:02,,1,m,VERTICAL_PROFILE,27.46,degC,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Barometric Pressure,5/20/2019 8:33:03,,2.005,m,VERTICAL_PROFILE,760.4,mmHg,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Concentration),5/20/2019 8:33:03,,2.005,m,VERTICAL_PROFILE,0.34,mg/l,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Saturation),5/20/2019 8:33:03,,2.005,m,VERTICAL_PROFILE,4.2,%,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,pH,5/20/2019 8:33:03,,2.005,m,VERTICAL_PROFILE,6.52,s.u.,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Specific Conductivity,5/20/2019 8:33:03,,2.005,m,VERTICAL_PROFILE,122,uS/cm,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Water Temperature_C,5/20/2019 8:33:03,,2.005,m,VERTICAL_PROFILE,25.96,degC,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Barometric Pressure,5/20/2019 8:33:51,,3,m,VERTICAL_PROFILE,760.4,mmHg,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Concentration),5/20/2019 8:33:51,,3,m,VERTICAL_PROFILE,0.32,mg/l,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Saturation),5/20/2019 8:33:51,,3,m,VERTICAL_PROFILE,3.7,%,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,pH,5/20/2019 8:33:51,,3,m,VERTICAL_PROFILE,6.35,s.u.,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Specific Conductivity,5/20/2019 8:33:51,,3,m,VERTICAL_PROFILE,128,uS/cm,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Water Temperature_C,5/20/2019 8:33:51,,3,m,VERTICAL_PROFILE,21.49,degC,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Barometric Pressure,5/20/2019 8:34:44,,4.002,m,VERTICAL_PROFILE,760.3,mmHg,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Concentration),5/20/2019 8:34:44,,4.002,m,VERTICAL_PROFILE,0.31,mg/l,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Saturation),5/20/2019 8:34:44,,4.002,m,VERTICAL_PROFILE,3.3,%,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,pH,5/20/2019 8:34:44,,4.002,m,VERTICAL_PROFILE,6.23,s.u.,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Specific Conductivity,5/20/2019 8:34:44,,4.002,m,VERTICAL_PROFILE,146,uS/cm,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Water Temperature_C,5/20/2019 8:34:44,,4.002,m,VERTICAL_PROFILE,18.61,degC,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Barometric Pressure,5/20/2019 8:35:30,,5.006,m,VERTICAL_PROFILE,760.4,mmHg,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Concentration),5/20/2019 8:35:30,,5.006,m,VERTICAL_PROFILE,0.3,mg/l,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Saturation),5/20/2019 8:35:30,,5.006,m,VERTICAL_PROFILE,3.2,%,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,pH,5/20/2019 8:35:30,,5.006,m,VERTICAL_PROFILE,6.19,s.u.,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Specific Conductivity,5/20/2019 8:35:30,,5.006,m,VERTICAL_PROFILE,151,uS/cm,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Water Temperature_C,5/20/2019 8:35:30,,5.006,m,VERTICAL_PROFILE,17.66,degC,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Barometric Pressure,5/20/2019 8:36:44,,5.996,m,VERTICAL_PROFILE,760.4,mmHg,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Concentration),5/20/2019 8:36:44,,5.996,m,VERTICAL_PROFILE,0.28,mg/l,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,DO (Saturation),5/20/2019 8:36:44,,5.996,m,VERTICAL_PROFILE,3,%,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,pH,5/20/2019 8:36:44,,5.996,m,VERTICAL_PROFILE,6.07,s.u.,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Specific Conductivity,5/20/2019 8:36:44,,5.996,m,VERTICAL_PROFILE,216,uS/cm,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
+,Cn,Water Temperature_C,5/20/2019 8:36:44,,5.996,m,VERTICAL_PROFILE,17.54,degC,Preliminary,OK,WATER,Cn20190520,GRAB,,My_DeviceType,,,,,,,,,,,,,,,,,
