+++ Samples/DotNetSdk/SondeFileImporter/ExampleDataFiles/Sonde file example.csv
@@ -1,0 +1,9 @@
+Date,Time,Site,Unit ID,User ID,°C,mmHg,DO %,DO mg/L,SPC-uS/cm,pH,pH mV,DEP m,Batt V,Lat,Lon,Date Time,Sample ID
+5/20/2019,8:30:08,Cn,,JBG,27.57,760.4,51.6,4.07,124,7.25,-61.6,0.261,0,28.56602,-81.344,5/20/19 8:30,Cn20190520
+5/20/2019,8:31:07,Cn,,JBG,27.54,760.5,50.3,3.97,124,7.03,-48.5,0.497,0,28.56604,-81.34399,5/20/19 8:31,Cn20190520
+5/20/2019,8:32:02,Cn,,JBG,27.46,760.4,46.2,3.65,124,6.83,-37.2,1,0,28.56602,-81.34398,5/20/19 8:32,Cn20190520
+5/20/2019,8:33:03,Cn,,JBG,25.96,760.4,4.2,0.34,122,6.52,-19,2.005,0,28.56603,-81.34398,5/20/19 8:33,Cn20190520
+5/20/2019,8:33:51,Cn,,JBG,21.49,760.4,3.7,0.32,128,6.35,-9.3,3,0,28.56604,-81.344,5/20/19 8:33,Cn20190520
+5/20/2019,8:34:44,Cn,,JBG,18.61,760.3,3.3,0.31,146,6.23,-2.6,4.002,0,28.56603,-81.34401,5/20/19 8:34,Cn20190520
+5/20/2019,8:35:30,Cn,,JBG,17.66,760.4,3.2,0.3,151,6.19,-0.6,5.006,0,28.56603,-81.34401,5/20/19 8:35,Cn20190520
+5/20/2019,8:36:44,Cn,,JBG,17.54,760.4,3,0.28,216,6.07,5.9,5.996,0,28.56603,-81.34399,5/20/19 8:36,Cn20190520
