+++ Samples/DotNetSdk/SondeFileImporter/Properties/AssemblyInfo.cs
@@ -1,0 +1,7 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyTitle("Sonde File Importer")]
+[assembly: AssemblyDescription("Transforms sonde files to Samples observation csv file format and upload the file to Samples server.")]
+[assembly: AssemblyCompany("Aquatic Informatics Inc.")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]
