+++ Samples/DotNetSdk/SondeFileImporter/ServiceClient/ImportResultResponse.cs
@@ -1,8 +1,8 @@
-﻿using System;
+﻿
 
 // ReSharper disable InconsistentNaming
 
-namespace SondeFileSynchronizer.ServiceClient
+namespace SondeFileImporter.ServiceClient
 {
     public class ImportResultResponse
     {
