+++ Samples/DotNetSdk/SondeFileImporter/ServiceClient/ImportStatus.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Net;
 
-namespace SondeFileSynchronizer.ServiceClient
+namespace SondeFileImporter.ServiceClient
 {
     public class ImportStatus
     {
