+++ Samples/DotNetSdk/SondeFileImporter/packages.config
