--- Samples/DotNetSdk/SondeFileSynchronizer/Properties/AssemblyInfo.cs
@@ -1,7 +1,0 @@
-﻿using System.Reflection;
-
-[assembly: AssemblyTitle("Sonde File Samples Synchronizer")]
-[assembly: AssemblyDescription("Transforms sonde files to Samples data file format and upload the file to Samples server.")]
-[assembly: AssemblyCompany("Aquatic Informatics Inc.")]
-[assembly: AssemblyVersion("1.0.0.0")]
-[assembly: AssemblyFileVersion("1.0.0.0")]
